+++ src/SettingsList.h
@@ -645,6 +645,8 @@ inline const std::vector<SettingInfo>& getBaseSettingsList() {
                            "screenMarginHorizontal", StrId::STR_CAT_READER));
     add(SettingInfo::Toggle(StrId::STR_PUBLISHER_PAGE_NUMBERS, &CrossPointSettings::publisherPageNumbers,
                             "publisherPageNumbers", StrId::STR_CAT_READER));
+    add(SettingInfo::Value(StrId::STR_PROGRESS_DECIMAL_PLACES, &CrossPointSettings::progressPercentDecimalPlaces,
+                           {0, 2, 1}, "progressPercentDecimalPlaces", StrId::STR_CAT_READER));
     add(SettingInfo::Enum(
         StrId::STR_PARA_ALIGNMENT, &CrossPointSettings::paragraphAlignment,
         {StrId::STR_JUSTIFY, StrId::STR_ALIGN_LEFT, StrId::STR_CENTER, StrId::STR_ALIGN_RIGHT, StrId::STR_BOOK_S_STYLE},